Moved All qouries i could find releavent to Group #7
