+++ CheckInSystem/Models/Group.cs
@@ -13,7 +13,7 @@ public class Group : INotifyPropertyChanged
 {
     DatabaseHelper databaseHelper;
 
-    public int ID { get; private set; }
+    public int ID { get; set; }
     
     private string _name;
     public string Name
@@ -29,55 +29,16 @@ public class Group : INotifyPropertyChanged
         set => SetProperty(ref _isvisible, value);
     }
     
-    public ObservableCollection<Employee> Members { get; private set; }
+    public ObservableCollection<Employee> Members { get; set; }
 
     public static List<Group> GetAllGroups(List<Employee> employees)
     {
-        string selectQueryGroups = @"SELECT * FROM [group]";
-        string selectQueryEmployeesInGroup = @"SELECT employeeID FROM employeeGroup WHERE groupID = {0}";
-
-        using var connection = Database.Database.GetConnection();
-        if (connection == null)
-            throw new Exception("Could not establish database connection!");
-
-        var groups = connection.Query<Group>(selectQueryGroups).ToList();
-
-        foreach (var group in groups)
-        {
-            string formattedQuery = string.Format(selectQueryEmployeesInGroup, group.ID);
-            var employeeIDs = connection.Query<int>(formattedQuery).ToList();
-            group.Members = new ObservableCollection<Employee>();
-
-            foreach (var employeeID in employeeIDs)
-            {
-                var temp = employees.Where(i => i.ID == employeeID).FirstOrDefault();
-                if (temp != null)
-                {
-                    group.Members.Add(temp);
-                }
-            }
-        }
-
-        return groups;
+        return DatabaseHelper.GetAllGroups(employees);
     }
 
     public static Group NewGroup(String name)
     {
-        string insertQuery = @"INSERT INTO [group] (name) VALUES (@name)
-            SELECT SCOPE_IDENTITY();";
-        Group newGroup = new Group()
-        {
-            Name = name,
-            Members = new ObservableCollection<Employee>()
-        };
-
-        using var connection = Database.Database.GetConnection();
-        if (connection == null)
-            throw new Exception("Could not establish database connection!");
-
-        newGroup.ID = connection.QueryFirst<int>(insertQuery, new {name = name});
-
-        return newGroup;
+        return DatabaseHelper.NewGroup(name);
     }
 
     public void RemoveGroupDb()
@@ -97,32 +58,14 @@ public class Group : INotifyPropertyChanged
     
     public void AddEmployee(Employee employee)
     {
-        if (Members.Contains(employee)) return;
-        
-        string insertQuery = @"INSERT INTO employeeGroup (employeeID, groupID) VALUES (@employeeID, @groupID)";
-
-        using var connection = Database.Database.GetConnection();
-        if (connection == null)
-            throw new Exception("Could not establish database connection!");
-
-        connection.Query(insertQuery, new { employeeID = employee.ID, @groupID = this.ID });
-        
-        this.Members.Add(employee);
+        databaseHelper.AddEmployee(employee, Members, ID);
+        Members.Add(employee);
     }
 
     public void RemoveEmployee(Employee employee)
     {
-        if (!Members.Contains(employee)) return;
-        
-        string deletionQuery = @"DELETE employeeGroup WHERE employeeID = @employeeID AND groupID = @groupID";
-
-        using var connection = Database.Database.GetConnection();
-        if (connection == null)
-            throw new Exception("Could not establish database connection!");
-
-        connection.Query(deletionQuery, new { employeeID = employee.ID, @groupID = this.ID });
-
-        this.Members.Remove(employee);
+        databaseHelper.RemoveEmployee(employee, Members, ID);
+        Members.Remove(employee);
     }
 
     public event PropertyChangedEventHandler? PropertyChanged;
